// src/front/src/api/answerrule/models/timeSlotDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX AnswerRule API (via Gateway)
 * OpenAPI spec version: 1.0
 */
import type { DayOfWeek } from "./dayOfWeek";

export interface TimeSlotDto {
	/** @nullable */
	gid?: string | null;
	dayOfWeek?: DayOfWeek;
	/** @nullable */
	startTime?: string | null;
	/** @nullable */
	endTime?: string | null;
	isAllDay?: boolean;
}